python_api/requirements.txt
scholarly==1.7.11
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
pgvector==0.2.3
python-dotenv==1.0.0
openai==1.3.5
tqdm==4.66.1
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.2 